Refactored unroll prereqs
The map from k->rule used to determine recursive and non-recursive
KClauses is now precomputed. Informal testing indicates that this is
noticably faster.

Renamed and reorganized acessors for readability.

Moved large functions in `where` clauses to top-level, for readbility,
and because it makes it easier for `ghc` to optimize. Additionally, I
tried to add some consistency to the use of let/where.