Don't crash on unsound unrolling
When we recieve responses after interpolation, we expect them to be
interpolants, but this is only true if we give Z3 a valid interpolation
problem.

This can happen because unrolling is unsound in general, and in this
particular case could've given us a Horn formula corresponding to an
ill-typed program

In this case, Z3 simply tells us `Sat`, so we should return an empty
list of interpolants; give up on this unrolling and proceed from depth
`n` to depth `n+1`.