Cleanup Interpolate.hs
Take all linter solutions, clean up the types a little bit, moved some
code around to make it feel more organized.

Might be worth it to make a Solver/Interpolate/ folder and split this
600 line file into modules after PLDI, but hopefully that won't be
necessary --- should be able to unify treatment with Graph.hs, but
that'll take a lot of work.