Rearchiteching synthesize
Inside of `Sol.solve`, where we call `solve_` is actually a pretty good
place to do our business; we have both a `SolverInfo` and a `Result`, so
we know what kvars to solve for, how to get `cSucc` boiz, and where
we're starting from. It's even in the IO monad directly already, so we
won't have to do too much fucking with shit.

Let's start with the damn simple `bool00.fq` and see if we can
``synthesize'' the simple *conjuntive* invariant before we get all fancy
with condition abduction 'n' shiz.

> Wait, that's sort of dumb... we're already going to have a conjuctive
solution coming in. Throw it away? Let's throw it away.

No, that still doesn't work. Let's pretend that constraint `[3]` is the
failing one for now. We can actually force it to do that by *actually*
failing (by commenting out the quals), and thereby test all the code
leading up to the actual synthesis before that. Then, when we have that
working, we can comment out the code that figures out what cons to start
on and the old sol, hardcode it, and write synthesis on the conjuctive
boi.

TODO: this won't (at all) test the path where the code extracts
solutions.